fix #162 and #170: use entry point index as bit
This fixes a crash that happens when there are more than 8 entry points. The problem is due to some code that is meant to eventually be a part of the code splitting feature.

For code splitting, files are classified into separate chunks by which entry points they are reachable by. Reachability is determined using a bit set where the bit is identified as the index of the entry point in the entry points array.

Unfortunately I confused this index with the source index of the entry point, which is the index of the entry point in the overall array of sources, and which is much bigger. This causes a crash if there's an entry point with a source index of 8 or higher, since that tries to set a bit in the second byte of the bit set, which was never allocated because there's only ever one entry point in the current independent compilation mode (i.e. with code splitting disabled).